Incorporate feedback + refactoring code
